# [Fix/new feature] Add max_ram_usage_in_bytes parameter for better control of ram usage