--- a/embedding_reader/parquet_reader.py
+++ b/embedding_reader/parquet_reader.py
@@ -69,7 +69,7 @@ class ParquetReader:
         self.nb_files = len(self.headers["count"])
         self.byte_per_item = 4 * self.dimension
         self.total_size = self.count * self.byte_per_item
-        self.average_file_size = self.total_size / self.nb_files
+        self.max_file_size = max(self.headers["count"]) * self.byte_per_item
 
     def __call__(
         self,
@@ -95,7 +95,7 @@ class ParquetReader:
 
         if parallel_pieces is None:
             # read_piece function can load the whole file in memory to take only max_piece_size bytes
-            read_piece_ram_usage = self.average_file_size
+            read_piece_ram_usage = self.max_file_size
             # We try to parallelize a maximum as long at it fits the ram constraint
             parallel_pieces = min(max(int(math.ceil(max_ram_usage_in_bytes / read_piece_ram_usage)), 1), 50)
         pieces = build_pieces(