--- a/examples/exemple_parquet.py
+++ b/examples/exemple_parquet.py
@@ -2,10 +2,9 @@ from embedding_reader import EmbeddingReader
 
 print("Starting")
 embedding_reader = EmbeddingReader(
-    embeddings_folder="https://mystic.the-eye.eu/public/AI/cah/laion5b/embeddings/laion1B-nolang/img_emb/",
-    metadata_folder="https://mystic.the-eye.eu/public/AI/cah/laion5b/embeddings/laion1B-nolang/laion1B-nolang-metadata/",
-    meta_columns=["url", "caption"],
-    file_format="parquet_npy",
+    embeddings_folder="hdfs://root/user/cailimage/prod/image-text-pipeline/EU/prediction/clip/embeddings/with_image=true",
+    embedding_column="image_embs",
+    file_format="parquet",
 )
 print("embedding count", embedding_reader.count)
 print("dimension", embedding_reader.dimension)
@@ -16,5 +15,4 @@ print("byte per item", embedding_reader.byte_per_item)
 for emb, meta in embedding_reader(
     batch_size=10**6, start=0, end=embedding_reader.count, max_ram_usage_in_bytes=2**30
 ):
-    print(emb.shape)
-    print(meta["url"], meta["caption"])
+    pass