When reading a batch of embeddings of 24GB, the program memory usage could reach 125GB at peak time for 250MB embedding files.
With this commit, we can now control the ram usage + We never use more ram than necessary when the embedding files are big.

- Modification 1:
Before:
Now: max_piece_size = max(number of embedding for 50MB, batch size in MB).
Indeed, we can't memory-map with fsspec on distant files (https://github.com/fsspec/filesystem_spec/issues/832), and we can't load an exact part of data without loading the whole file.
The read_piece function would load 5 times a 250MB files to extract the 5 pieces of 50MB it contains, so it is better to set max_piece_size = batch_size when batch_size > max_piece_size

- Modification 2:
Before: parallel_pieces = max(math.ceil(batch_size / max_piece_size), 10)
Now: parallel_pieces = min(round(max_ram_usage_in_bytes/max_piece_size), 50).
For a batch requiring 600 pieces, we would have needed to read 600 files in the meantime! So we cap the limit to 50 max