@rom1504 2022-07-07
I think doing https://github.com/rom1504/embedding-reader/pull/25 for the parquet reader will improve memory usage significantly