@victor-paltz 2022-07-08
> I think doing #25 for the parquet reader will improve memory usage significantly

Ah nice! This PR avoids to load twice the same file if two pieces are in the same file 😎 

If we were unlucky though, we would still have bad memory usage if all the pieces were coming from a different file (even if it is less likely). So I think the best thing would still be to increase the piece size to be bigger than the maximum number of embeddings of the biggest file. WDYT?

What is the advantage of having _piece size < max file size_? By taking bigger pieces we wouldn't even need to bother to read twice the same file. Is it because with memory mapping on disk the cost is zero and loading a huge file would have broken things?