@rom1504 2022-07-08
for the numpy format and files that are remote, range requests in http (and similar in s3 and hdfs) allow to get many parts of the file in parallel which makes reading much faster
for parquet indeed there is much less advantage indeed, as the smallest part that can actually be requests is a row group

https://github.com/rom1504/embedding-reader/pull/25 makes it possible to have parquet still be reasonably fast, but maybe a sequential reader for parquet would be just as fast, I'm not sure, would be interesting to benchmark

anyway I will check your PR