@rom1504 2022-07-10
can you check if this is as fast as before ? (for example by running https://github.com/rom1504/embedding-reader/blob/main/examples/example_numpy_parquet.py ?)