@victor-paltz 2022-07-21
> can you check if this is as fast as before ? (for example by running https://github.com/rom1504/embedding-reader/blob/main/examples/example_numpy_parquet.py ?)

I run the code on the example_parquet.py (so only the parquet reader), here are my results:

### Configuration:
- batch of 10^6
- 16800 files
- file size 220MB
- embedding size 512d

### Previous version: (git checkout 8630c739765558692ce73773cc189036c7c7819f)
Nb pieces: 101000
- I needed to lower the number of threads because there were 500 and each one was reopening a file
- With 50 threads: I get killed at one point and the average time was 30h to read through the dataset
- With 10 threads: I get killed at one point and the average time was 38h to read through the dataset
- With 1   thread:   The average time was +100h to read through the dataset

### New version:
_parquet reader with max_piece_size = max(50MB, batch_size, 1)_
    - ~28s to load batches of size (1000000, 512), total time: ~16h  -> 18819 pieces     (parallel_pieces=50)
    - ~3s  to load batches of size (100000, 512),  total time: ~14h  -> 37482 pieces       (parallel_pieces=50)
_parquet reader with max_piece_size = max(50MB, 1), max ram usage = 4GB_
    - ~28s to load batches of size (1000000, 512), total time: ~15h  -> 100316 pieces     (parallel_pieces=50)
    - ~3s  to load batches of size (100000, 512),  total time: ~14h  -> 105635 pieces       (parallel_pieces=50)
_parquet reader with max_piece_size = max(50MB, 1), max ram usage = 1GB_
    - ~28s to load batches of size (1000000, 512), total time: ~15h  -> 100316 pieces     (parallel_pieces=21 or 10)
    - ~3s  to load batches of size (100000, 512),  total time: ~14h  -> 105635 pieces       (parallel_pieces=21 or 10)
-> No impact on increasing the number of pieces (maybe even slightly faster? but not significative)

### Conclusion
Forcing a piece size of batch_size doesn't give any uplift + is less efficient when using a numpy reader. As a consequence, it is better to keep the same code for everyone and do `max_piece size = 50MB`  over `max(50MB, batch_size)`.