@rom1504 2022-08-02
Sounds pretty good for the parquet reader. However I would like to check for numpy too as it has different performance characteristics (much faster)

I'll try and check that tonight.
If you have time, you can also run it, the numpy data in the example is public (and fsspec handles http protocol)