@rom1504 2022-08-07
sorry for the delay

so I'm trying this now.
For numpy+parquet it works just the same as before

for numpy, it seems to be getting stuck 
I'm running https://github.com/rom1504/embedding-reader/blob/main/examples/example_numpy_parquet.py

checking