@rom1504 2022-08-07
I think it would still be worth it to do https://github.com/rom1504/embedding-reader/issues/26